+++ packages/config-registry-controller/src/index.ts
@@ -16,7 +16,11 @@ export {
   ConfigRegistryController,
   DEFAULT_POLLING_INTERVAL,
 } from './ConfigRegistryController.js';
-export { selectFeaturedNetworks, selectNetworks } from './selectors.js';
+export {
+  selectFeaturedNetworks,
+  selectNetworks,
+  selectEvmAutoEnabledNetworksChainIds,
+} from './selectors.js';
 export type {
   FetchConfigOptions,
   FetchConfigResult,